refactor: move ResultsViewer to separate file in utils
